extract parts of the content script
